4.1.0→4.2.0 porting and f64 precision improvements
- Port spessasynth_core 4.1.0 to 4.2.0 changes (effects, delay, insertion, drum params, etc.)
- f64化 Phase 1: effects directory internal calculations
- f64化 Phase 2: tuning/filter parameters (voice.tuning_cents, lowpass fc_offset,
  set_modulation_depth/set_tuning/set_custom_controller params, get_tuning, data_entry casts)
- Result: 6/9 PASS songs now at Max diff=1 (near-perfect match with TypeScript)

+++ src/synthesizer/audio_engine/engine_components/dsp_chain/lowpass_filter.rs
@@ -169,7 +169,7 @@ impl LowpassFilter {
         sample_count: usize,
         modulated_generators: &[i16],
         output_buffer: &mut [f32],
-        fc_offset: f32,
+        fc_offset: f64,
     ) {
         // Use f64 for filter cutoff arithmetic to match TS (JS number is f64).
         let initial_fc = modulated_generators[gt::INITIAL_FILTER_FC as usize] as f64;
@@ -184,7 +184,7 @@ impl LowpassFilter {
             self.current_initial_fc = initial_fc;
         }
 
-        let target_cutoff = self.current_initial_fc + fc_offset as f64;
+        let target_cutoff = self.current_initial_fc + fc_offset;
         let modulated_resonance = modulated_generators[gt::INITIAL_FILTER_Q as usize];
 
         // Filter bypass: fully open + no resonance.
